refactor: Improve error handling and message formatting in interpreter and operation functions
- Introduced NewErrorFValue for formatted error messages.
- Updated various functions to use formatted error messages for better clarity.
- Adjusted error handling in array operations and function calls to provide more informative feedback.

+++ src/global.h
@@ -497,7 +497,7 @@ typedef struct user_function_struct {
     Value*       Scope;        /**< The environment where the function was defined */
     String       Name;         /**< Function name (Nullable) */
     uint8_t*     Codes;        /**< Bytecode instructions */
-    int          CodeC;        /**< Size of bytecode */
+    size_t       CodeC;        /**< Size of bytecode */
     int          Argc;         /**< Argument count */
     int          LocalC;       /**< Local variable count */
     CaptureMeta* CaptureMetas; /**< Array of capture metadata */
